+++ src/main/java/it/bitify/libreria/repository/LibroRepo.java
@@ -3,5 +3,5 @@ package it.bitify.libreria.repository;
 import it.bitify.libreria.entity.Libro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LibroRepo extends JpaRepository<Libro, Integer> {
+public interface LibroRepo extends JpaRepository<Libro, Long> {
 }
