+++ src/main/java/it/bitify/libreria/repository/PrestitoRepo.java
@@ -3,5 +3,5 @@ package it.bitify.libreria.repository;
 import it.bitify.libreria.entity.Prestito;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PrestitoRepo extends JpaRepository<Prestito, Integer> {
+public interface PrestitoRepo extends JpaRepository<Prestito, Long> {
 }
